requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.8.2
redis==5.0.8
